+++ sparksteps/defaults.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Application CLI defaults
+"""
+DEFAULT_JOBFLOW_ROLE = 'EMR_EC2_DefaultRole'
+DEFAULT_APP_LIST = ['Hadoop', 'Spark']