+++ tests/test_sparksteps.py
@@ -6,7 +6,6 @@ import os.path
 import boto3
 import moto
 
-from sparksteps import __main__
 from sparksteps.cluster import emr_config
 from sparksteps.steps import setup_steps, S3DistCp
 
@@ -24,6 +23,7 @@ EPISODES_AVRO = os.path.join(DATA_DIR, 'episodes.avro')
 def test_emr_cluster_config():
     config = emr_config('emr-5.2.0',
                         instance_type_master='m4.large',
+                        jobflow_role='MyCustomRole',
                         keep_alive=False,
                         instance_type_core='m4.2xlarge',
                         instance_type_task='m4.2xlarge',
@@ -31,7 +31,8 @@ def test_emr_cluster_config():
                         num_task=1,
                         bid_price_task='0.1',
                         maximize_resource_allocation=True,
-                        name="Test SparkSteps")
+                        name="Test SparkSteps",
+                        app_list=['hadoop', 'hive', 'spark'])
     assert config == {'Instances':
                           {'InstanceGroups': [{'InstanceCount': 1,  # NOQA: E127
                                                'InstanceRole': 'MASTER',
@@ -52,9 +53,9 @@ def test_emr_cluster_config():
                            'KeepJobFlowAliveWhenNoSteps': False,
                            'TerminationProtected': False
                            },
-                      'Applications': [{'Name': 'Hadoop'}, {'Name': 'Spark'}],
+                      'Applications': [{'Name': 'Hadoop'}, {'Name': 'Hive'}, {'Name': 'Spark'}],
                       'Name': 'Test SparkSteps',
-                      'JobFlowRole': 'EMR_EC2_DefaultRole',
+                      'JobFlowRole': 'MyCustomRole',
                       'ReleaseLabel': 'emr-5.2.0',
                       'VisibleToAllUsers': True,
                       'ServiceRole': 'EMR_DefaultRole',
@@ -343,99 +344,3 @@ def test_s3_dist_cp_step():
             'Jar': 'command-runner.jar'},
         'Name': 'S3DistCp step'
     }
-
-
-def test_parser():
-    parser = __main__.create_parser()
-    cmd_args_str = """episodes.py \
-      --s3-bucket my-bucket \
-      --aws-region us-east-1 \
-      --release-label emr-4.7.0 \
-      --uploads examples/dir examples/episodes.avro \
-      --submit-args="--jars /home/hadoop/lib/spark-avro_2.10-2.0.2.jar" \
-      --app-args="--input /home/hadoop/episodes.avro" \
-      --num-core 1 \
-      --tags Name=MyName CostCenter=MyCostCenter \
-      --defaults spark-defaults key=value another_key=another_value \
-      --maximize-resource-allocation \
-      --debug
-    """
-    args = __main__.parse_cli_args(parser, args=shlex.split(cmd_args_str))
-    assert args['app'] == 'episodes.py'
-    assert args['s3_bucket'] == 'my-bucket'
-    assert args['app_args'] == ['--input', '/home/hadoop/episodes.avro']
-    assert args['debug'] is True
-    assert args['defaults'] == ['spark-defaults', 'key=value', 'another_key=another_value']
-    assert args['instance_type_master'] == 'm4.large'
-    assert args['release_label'] == 'emr-4.7.0'
-    assert args['submit_args'] == ['--jars',
-                                   '/home/hadoop/lib/spark-avro_2.10-2.0.2.jar']
-    assert args['uploads'] == ['examples/dir', 'examples/episodes.avro']
-    assert args['tags'] == ['Name=MyName', 'CostCenter=MyCostCenter']
-    assert args['maximize_resource_allocation'] is True
-    assert args['num_core'] == 1
-
-
-def test_parser_deprecated_args():
-    parser = __main__.create_parser()
-    cmd_args_str = """episodes.py \
-      --s3-bucket my-bucket \
-      --aws-region us-east-1 \
-      --release-label emr-4.7.0 \
-      --uploads examples/dir examples/episodes.avro \
-      --submit-args="--jars /home/hadoop/lib/spark-avro_2.10-2.0.2.jar" \
-      --app-args="--input /home/hadoop/episodes.avro" \
-      --master m4.4xlarge \
-      --slave c3.8xlarge \
-      --num-core 1 \
-      --dynamic-pricing \
-      --tags Name=MyName CostCenter=MyCostCenter \
-      --defaults spark-defaults key=value another_key=another_value \
-      --maximize-resource-allocation \
-      --debug
-    """
-    args = __main__.parse_cli_args(parser, args=shlex.split(cmd_args_str))
-    assert args['app'] == 'episodes.py'
-    assert args['s3_bucket'] == 'my-bucket'
-    assert args['app_args'] == ['--input', '/home/hadoop/episodes.avro']
-    assert args['debug'] is True
-    assert args['defaults'] == ['spark-defaults', 'key=value', 'another_key=another_value']
-    assert args['instance_type_master'] == 'm4.4xlarge'
-    assert args['instance_type_core'] == 'c3.8xlarge'
-    assert args['dynamic_pricing_task'] is True
-    assert args['release_label'] == 'emr-4.7.0'
-    assert args['submit_args'] == ['--jars',
-                                   '/home/hadoop/lib/spark-avro_2.10-2.0.2.jar']
-    assert args['uploads'] == ['examples/dir', 'examples/episodes.avro']
-    assert args['tags'] == ['Name=MyName', 'CostCenter=MyCostCenter']
-    assert args['maximize_resource_allocation'] is True
-
-
-def test_parser_with_bootstrap():
-    parser = __main__.create_parser()
-    cmd_args_str = """episodes.py \
-      --s3-bucket my-bucket \
-      --aws-region us-east-1 \
-      --release-label emr-4.7.0 \
-      --uploads examples/dir examples/episodes.avro \
-      --submit-args="--jars /home/hadoop/lib/spark-avro_2.10-2.0.2.jar" \
-      --app-args="--input /home/hadoop/episodes.avro" \
-      --num-core 1 \
-      --tags Name=MyName CostCenter=MyCostCenter \
-      --defaults spark-defaults key=value another_key=another_value \
-      --bootstrap-script s3://bucket/bootstrap-actions.sh \
-      --debug
-    """
-    args = __main__.parse_cli_args(parser, args=shlex.split(cmd_args_str))
-    assert args['app'] == 'episodes.py'
-    assert args['s3_bucket'] == 'my-bucket'
-    assert args['app_args'] == ['--input', '/home/hadoop/episodes.avro']
-    assert args['debug'] is True
-    assert args['defaults'] == ['spark-defaults', 'key=value', 'another_key=another_value']
-    assert args['instance_type_master'] == 'm4.large'
-    assert args['release_label'] == 'emr-4.7.0'
-    assert args['submit_args'] == ['--jars',
-                                   '/home/hadoop/lib/spark-avro_2.10-2.0.2.jar']
-    assert args['uploads'] == ['examples/dir', 'examples/episodes.avro']
-    assert args['tags'] == ['Name=MyName', 'CostCenter=MyCostCenter']
-    assert args['bootstrap_script'] == 's3://bucket/bootstrap-actions.sh'