+++ CHANGELOG.rst
@@ -6,6 +6,13 @@ Changelog
 Releases
 --------
 
+v1.1.0 (2019-07-13)
+~~~~~~~~~~~~~~~~~~~
+
+* Added `jobflow_role` CLI argument to configure cluster EC2 Instance Profile
+* Added `app-list` CLI argument to configure list of Applications installed on cluster
+
+
 v1.0.0 (2019-07-03)
 ~~~~~~~~~~~~~~~~~~~